# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dataset Studio Engine",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/projects/{project_id}": {
      "get": {
        "tags": [],
        "operationId": "project",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/artifacts": {
      "get": {
        "tags": [
          "tools"
        ],
        "operationId": "artifacts",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Artifacts"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/projects/{project_id}/sources/{source_id}/assets/{asset_id}/metadata": {
      "get": {
        "tags": [],
        "operationId": "metadata",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetadataOverview"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        }
      },
      "Artifact": {
        "type": "object",
        "required": [
          "id",
          "project_id",
          "job_id",
          "output_id",
          "name",
          "kind",
          "schema_version",
          "state",
          "created_at",
          "files",
          "provenance"
        ],
        "properties": {
          "count": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "created_at": {
            "type": "string"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtifactFile"
            }
          },
          "id": {
            "type": "string"
          },
          "issue": {
            "type": [
              "string",
              "null"
            ]
          },
          "job_id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "output_id": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ArtifactProvenance"
          },
          "schema_version": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "state": {
            "$ref": "#/components/schemas/ArtifactState"
          }
        }
      },
      "ArtifactFile": {
        "type": "object",
        "required": [
          "path",
          "media_type"
        ],
        "properties": {
          "bytes": {
            "type": [
              "string",
              "null"
            ]
          },
          "media_type": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "sha256": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ArtifactProvenance": {
        "type": "object",
        "required": [
          "input_artifacts",
          "fields_frozen",
          "evidence"
        ],
        "properties": {
          "attempt": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "evidence": {
            "type": "string"
          },
          "fields_frozen": {
            "type": "boolean"
          },
          "input_artifacts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "input_scope": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ScopeRef"
              }
            ]
          },
          "input_sha256": {
            "type": [
              "string",
              "null"
            ]
          },
          "run": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OperatorRun"
              }
            ]
          }
        }
      },
...
      "ArtifactState": {
        "type": "string",
        "enum": [
          "legacy",
          "publishing",
          "ready",
          "released",
          "unavailable"
        ]
      },
      "Artifacts": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Artifact"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "AssetKey": {
        "type": "object",
        "required": [
          "source_id",
          "asset_id"
        ],
        "properties": {
          "asset_id": {
            "type": "string"
          },
          "source_id": {
            "type": "string"
          }
        }
      },
...
      "AssetRecord": {
        "type": "object",
        "required": [
          "record_id"
        ],
        "properties": {
          "origin_observation_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "post_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "record_id": {
            "type": "string"
          },
          "source_md5": {
            "type": [
              "string",
              "null"
            ]
          },
          "storage_profile": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "MetadataObject": {
        "type": "object",
        "required": [
          "key",
          "name",
          "bytes",
          "extension",
          "source_name"
        ],
        "properties": {
          "bytes": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          },
          "key": {
            "$ref": "#/components/schemas/AssetKey"
          },
          "name": {
            "type": "string"
          },
          "source_name": {
            "type": "string"
          }
        }
      },
      "MetadataOverview": {
        "type": "object",
        "required": [
          "object",
          "dimensions_evidence",
          "records",
          "version"
        ],
        "properties": {
          "dimensions_evidence": {
            "type": "string"
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "object": {
            "$ref": "#/components/schemas/MetadataObject"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetRecord"
            }
          },
          "stored_height": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "stored_width": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "version": {
            "$ref": "#/components/schemas/ReadVersion"
          }
        }
      },
...
      "OperatorRun": {
        "type": "object",
        "required": [
          "operator_id",
          "operator_version",
          "parameters_version",
          "parameters"
        ],
        "properties": {
          "operator_id": {
            "type": "string"
          },
          "operator_version": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "parameters": {},
          "parameters_version": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
...
      "Project": {
        "type": "object",
        "required": [
          "id",
          "name",
          "directory",
          "created_at",
          "revision"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "directory": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "revision": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "ReadVersion": {
        "type": "object",
        "required": [
          "token",
          "library_id",
          "generation",
          "catalog_sequence",
          "analysis_sequence",
          "consistency"
        ],
        "properties": {
          "analysis_sequence": {
            "type": "string"
          },
          "catalog_sequence": {
            "type": "string"
          },
          "consistency": {
            "type": "string"
          },
          "generation": {
            "type": "string"
          },
          "library_id": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
...
      "ScopeRef": {
        "type": "object",
        "required": [
          "project_id",
          "target"
        ],
        "properties": {
          "project_id": {
            "type": "string"
          },
          "target": {
            "$ref": "#/components/schemas/ScopeTarget"
          }
        },
        "additionalProperties": false
      },
      "ScopeTarget": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "source_id",
              "revision",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "source"
                ]
              },
              "revision": {
                "type": "string"
              },
              "source_id": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "collection_id",
              "kind"
            ],
            "properties": {
              "collection_id": {
                "type": "string"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "workset"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "result_id",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "query_result"
                ]
              },
              "result_id": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "revision",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "selection"
                ]
              },
              "revision": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              }
            }
          }
        ]
      },
...
    }
  }
}